Move save_metadata function.
`timeline.rs` seems like a better home for it.
